+++ packages/cinterop/src/commonMain/kotlin/io/realm/kotlin/internal/interop/MemAllocator.kt
@@ -32,6 +32,11 @@ interface MemAllocator {
     fun allocRealmValueT(): RealmValueT
 
     /**
+     * Allocates a contiguous list of `realm_value_t` structs.
+     */
+    fun allocRealmValueList(count: Int): RealmValueList
+
+    /**
      * Instantiates a [RealmValue] representing a `realm_value_t` containing `null`.
      */
     // TODO optimize: investigate if we can statically create a null transport and reuse it
@@ -90,10 +95,10 @@ interface MemAllocator {
     fun realmObjectTransport(value: RealmObjectInterop?): RealmValue
 
     /**
-     * Instantiates a [RealmQueryArgsTransport] representing a `realm_query_arg_t` which in turn
-     * contains one or more `realm_value_t` as arguments.
+     * Instantiates a [RealmQueryArgumentList] representing a `realm_query_arg_t` that describe and
+     * references the incoming [RealmValueList] arguments.
      */
-    fun queryArgsOf(queryArgs: Array<RealmValue>): RealmQueryArgsTransport
+    fun queryArgsOf(queryArgs: List<RealmValueList>): RealmQueryArgumentList
 }
 
 /**