+++ packages/cinterop/src/jvm/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -1597,15 +1597,15 @@ actual object RealmInterop {
         realm: RealmPointer,
         classKey: ClassKey,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
         return LongPointerWrapper(
             realmc.realm_query_parse(
                 realm.cptr(),
                 classKey.key,
                 query,
-                args.first.toLong(),
-                args.second.value
+                args.size,
+                args.value
             )
         )
     }
@@ -1613,15 +1613,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_results(
         results: RealmResultsPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return LongPointerWrapper(
             realmc.realm_query_parse_for_results(
                 results.cptr(),
                 query,
-                count.toLong(),
-                args.second.value
+                args.size,
+                args.value
             )
         )
     }
@@ -1629,15 +1628,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_list(
         list: RealmListPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return LongPointerWrapper(
             realmc.realm_query_parse_for_list(
                 list.cptr(),
                 query,
-                count.toLong(),
-                args.second.value
+                args.size,
+                args.value
             )
         )
     }
@@ -1645,15 +1643,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_set(
         set: RealmSetPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return LongPointerWrapper(
             realmc.realm_query_parse_for_set(
                 set.cptr(),
                 query,
-                count.toLong(),
-                args.second.value
+                args.size,
+                args.value
             )
         )
     }
@@ -1684,15 +1681,10 @@ actual object RealmInterop {
     actual fun realm_query_append_query(
         query: RealmQueryPointer,
         filter: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
         return LongPointerWrapper(
-            realmc.realm_query_append_query(
-                query.cptr(),
-                filter,
-                args.first.toLong(),
-                args.second.value
-            )
+            realmc.realm_query_append_query(query.cptr(), filter, args.size, args.value)
         )
     }