+++ packages/cinterop/src/nativeDarwin/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -1303,15 +1303,15 @@ actual object RealmInterop {
         realm: RealmPointer,
         classKey: ClassKey,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
         return CPointerWrapper(
             realm_wrapper.realm_query_parse(
                 realm.cptr(),
                 classKey.key.toUInt(),
                 query,
-                args.first.toULong(),
-                args.second.value.ptr
+                args.size,
+                args.value.ptr
             )
         )
     }
@@ -1319,15 +1319,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_results(
         results: RealmResultsPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return CPointerWrapper(
             realm_wrapper.realm_query_parse_for_results(
                 results.cptr(),
                 query,
-                count.toULong(),
-                args.second.value.ptr
+                args.size,
+                args.value.ptr
             )
         )
     }
@@ -1335,15 +1334,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_list(
         list: RealmListPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return CPointerWrapper(
             realm_wrapper.realm_query_parse_for_list(
                 list.cptr(),
                 query,
-                count.toULong(),
-                args.second.value.ptr
+                args.size,
+                args.value.ptr
             )
         )
     }
@@ -1351,15 +1349,14 @@ actual object RealmInterop {
     actual fun realm_query_parse_for_set(
         set: RealmSetPointer,
         query: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
-        val count = args.first
         return CPointerWrapper(
             realm_wrapper.realm_query_parse_for_set(
                 set.cptr(),
                 query,
-                count.toULong(),
-                args.second.value.ptr
+                args.size,
+                args.value.ptr
             )
         )
     }
@@ -1400,14 +1397,14 @@ actual object RealmInterop {
     actual fun realm_query_append_query(
         query: RealmQueryPointer,
         filter: String,
-        args: Pair<Int, RealmQueryArgsTransport>
+        args: RealmQueryArgumentList
     ): RealmQueryPointer {
         return CPointerWrapper(
             realm_wrapper.realm_query_append_query(
                 query.cptr(),
                 filter,
-                args.first.toULong(),
-                args.second.value.ptr
+                args.size,
+                args.value.ptr
             )
         )
     }
@@ -3140,8 +3137,8 @@ actual object RealmInterop {
      */
     private fun Array<RealmValue>.toQueryArgs(memScope: MemScope): CPointer<realm_query_arg_t> {
         with(memScope) {
-            val cArgs = allocArray<realm_query_arg_t>(this@toQueryArgs.size)
-            this@toQueryArgs.mapIndexed { i, arg ->
+            val cArgs: CPointer<realm_query_arg_t> = allocArray<realm_query_arg_t>(this@toQueryArgs.size)
+            this@toQueryArgs.mapIndexed { i, arg: RealmValue ->
                 cArgs[i].apply {
                     this.nb_args = 1.toULong()
                     this.is_list = false